+++ CHANGELOG.md
@@ -2,6 +2,63 @@
 
 [Semantic Versioning](https://semver.org/)
 
+## [1.5.1] - 2021-04-30
+### Bug fixes
+- 'Membrane Orientation' preset now honors assembly-ids
+
+## [1.5.0] - 2021-04-22
+### Added
+- feature density preset for ligand validation
+
+### Bug fixes
+- hide 'Membrane Orientation' representation from UI
+
+## [1.4.2] - 2021-04-20
+### Bug fixes
+- fix measurement labels
+- hide 'Membrane Orientation' preset from UI
+
+## [1.4.1] - 2021-04-06
+### Bug fixes
+- structure export: getting transformed data
+
+## [1.4.0] - 2021-04-02
+### Added
+- membrane orientation preset by ANVIL
+
+## [1.3.5] - 2021-03-25
+### Bug fixes
+- make sure only structure objects are filtered in for export
+- do not create unit cell object when alignments are loaded
+
+## [1.3.4] - 2021-03-25
+### Bug fixes
+- structure export: getting children for a referenced node correctly
+
+## [1.3.3] - 2021-03-24
+### Bug fixes
+- build
+
+## [1.3.2] - 2021-03-24
+### General
+- major version update of Mol*
+
+## [1.3.1] - 2021-03-11
+### Bug fixes
+- this is a dummy release
+
+## [1.3.0] - 2021-03-10
+### Added
+- expose video export option
+
+## [1.2.3] - 2021-03-03
+### Bug fixes
+- warn when residues are more than 15 A apart (RO-2597)
+
+## [1.2.2] - 2021-02-26
+### Bug fixes
+- switch to 'model' when ligand preset is requested but ligand is not present in assembly '1' - HELP-16678
+
 ## [1.2.1] - 2021-02-18
 ### Bug fixes
 - limit number of exchanges per position